The-Eye/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langfuse>=3
python-dotenv>=1.0.0
ulid-py>=1.0.0,<2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
networkx>=3.0
